src/webhooks/receiver/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
mangum==0.19.0
boto3==1.36.26
python-dotenv==1.0.1
